+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change log
 
-## [v0.0.1](https://github.com/anotherbyte-net/gather-vision-gov-au-plugin/commits/v0.0.1)
+## [v0.0.1](https://github.com/anotherbyte-net/gather-vision/releases/tag/v0.0.1)
 
-*unreleased*
+[full change log](https://github.com/anotherbyte-net/gather-vision/commits/v0.0.1)
+
+- Initial implementation and placeholders.
+- Add basic plugin and tests.